major changes: termino de mergear los tps en mvc
