+++ back/microservices/AuthService/controllers/authController.js
@@ -1,19 +1,14 @@
 import { generateToken, refreshToken } from '../services/authService.js';
+//import { refreshToken as generateNewAccessToken } from '../services/authService.js';
 
-export const login = (req, res) => {
+export const loginController = (req, res) => {
   const { username, role } = req.user;
-
   const accessToken = generateToken({ username, role }, '1m');
-  const refreshTokenValue = generateToken({ username, role }, '15m');
-
-  res.json({
-    accessToken,
-    refreshToken: refreshTokenValue,
-    message: 'Autenticación exitosa',
-  });
+  const refreshToken = generateToken({ username, role }, '15m');
+  res.json({ accessToken, refreshToken, message: 'OK' });
 };
 
-export const refresh = async (req, res) => {
+export const refreshController = async (req, res) => {
   try {
     const newAccessToken = await refreshToken(req.body.refreshToken);
     res.json({ token: newAccessToken });
