+++ back/microservices/AuthService/middlewares/authMiddleware.js
@@ -42,26 +42,6 @@ const authenticateUser = (req, res, next) => {
   }
 };
 
-// Verificamosssss el token JWT
-const verifyToken = (req, res, next) => {
-  const authHeader = req.headers.authorization;
-
-  if (!authHeader || !authHeader.startsWith('Bearer ')) {
-    return res.status(401).json({ error: 'Acceso no autorizado: Token no proporcionado' });
-  }
-
-  const token = authHeader.split(' ')[1];
-
-  jwt.verify(token, JWT_SECRET, (err, decoded) => {
-    if (err) {
-      return res.status(401).json({ error: 'Acceso no autorizado: Token inválido o expirado' });
-    }
-    req.user = decoded;
-    next();
-  });
-};
 
 
-
-
-export { authMiddleware, authenticateUser, verifyToken };
+export { authMiddleware, authenticateUser };
